- Pr comments

+++ src/main/java/com/example/conduit/modules/article/repository/CommentRepository.java
@@ -6,6 +6,6 @@ import com.example.conduit.shared.GenericRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface CommentRepository extends GenericRepository<CommentData, Integer> {
+public interface CommentRepository extends GenericRepository<CommentData, Long> {
     Optional<List<CommentData>> getAllByArticle_Slug(String articleSlug);
 }
