+++ src/main/java/com/example/conduit/modules/user/repository/UserRepository.java
@@ -5,7 +5,7 @@ import com.example.conduit.shared.GenericRepository;
 
 import java.util.Optional;
 
-public interface  UserRepository extends GenericRepository<User, Integer> {
+public interface  UserRepository extends GenericRepository<User, Long> {
     Optional<User> findByEmail(String email);
     Optional<User> findByUsername(String username);
 }
